importccl: split up import helpers
Just moving code, no behavior changes.

Release note: none.
